refactor : Bean으로 주입받는 변수들을 final로 수정
