+++ src/main/java/wooteco/subway/admin/common/exception/InvalidSubwayPathException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.admin.path.domain;
+package wooteco.subway.admin.common.exception;
 
 public class InvalidSubwayPathException extends RuntimeException {
