+++ src/main/java/wooteco/subway/admin/path/controller/PathController.java
@@ -13,7 +13,7 @@ import wooteco.subway.admin.path.service.dto.PathInfoResponse;
 @RestController
 public class PathController {
 
-	private PathService pathService;
+	private final PathService pathService;
 
 	public PathController(PathService pathService) {
 		this.pathService = pathService;
